tools/facemotion/requirements.txt
ffmpeg-python==0.2.0
numpy==1.24.3
opencv-python==4.7.0.72
